requirements.txt
streamlit
Pillow
matplotlib
pandas
numpy
yfinance